+++ js/options.js
@@ -1,42 +1,29 @@
+const optionIds = ['jiraPath', 'pageRegex', 'regex', 'documentQueries'];
+
 function save() {
-  const pgOptions = {
-    jiraPath: document.getElementById('jiraPath').value,
-    pageRegex: document.getElementById('pageRegex').value,
-    regex: document.getElementById('regex').value
-  };
-
   const defaultOptions = getDefaultOptions();
-  if (!pgOptions.jiraPath.length) {
-    pgOptions.jiraPath = defaultOptions.jiraPath;
-  }
-  if (!pgOptions.pageRegex.length) {
-    pgOptions.pageRegex = defaultOptions.pageRegex;
-  }
-  if (!pgOptions.regex.length) {
-    pgOptions.regex = defaultOptions.regex;
-  }
+  const options = optionIds.reduce((acc, option) => {
+    acc[option] = document.getElementById(option).value || defaultOptions[option];
+    return acc;
+  }, {});
 
-  chrome.storage.sync.set({ options: JSON.stringify(pgOptions) });
+  chrome.storage.sync.set({ options });
 }
 
 function init() {
   chrome.storage.sync.get(['options'], (result) => {
-    const options = (result.options && JSON.parse(result.options)) || getDefaultOptions();
-
-    document.getElementById('jiraPath').value = options.jiraPath;
-    document.getElementById('pageRegex').value = options.pageRegex;
-    document.getElementById('regex').value = options.regex;
+    const defaults = getDefaultOptions();
+    const custom = result.options;
 
     document.getElementById('optionsForm').onchange = save;
 
-    document.getElementById('jiraPath').onkeyup = save;
-    document.getElementById('jiraPath').onclick = save;
-
-    document.getElementById('pageRegex').onkeyup = save;
-    document.getElementById('pageRegex').onclick = save;
-
-    document.getElementById('regex').onkeyup = save;
-    document.getElementById('regex').onclick = save;
+    optionIds.forEach((optionId) => {
+      const element = document.getElementById(optionId);
+      element.placeholder = custom[optionId] || defaults[optionId];
+      if (custom[optionId] !== defaults[optionId]) {
+        element.value = custom[optionId];
+      }
+    });
   });
 }